+++ src/components/board/BoardWrite.js
@@ -1,16 +1,16 @@
 import React, { useState, useEffect } from 'react';
-import { useParams } from 'react-router-dom';
+
 import { Button, Form, Row, Col } from 'react-bootstrap';
-import axios from "axios";
+
 import { useNavigate } from "react-router-dom";
 
 import 'bootstrap/dist/css/bootstrap.min.css';
 import '../../styles/board/boardWrite.scss';
 
 import BoardNav from './BoardNav';
-import BoardUploadFile from './BoardFileUpload/BoardFileUpload';
 
 import { postBoardWrite } from '../../api/BoardApi';
+// import BoardFileUpload from './BoardFileUpload/BoardFileUpload';
 
 function BoardWrite() {
     const navigate = useNavigate();
@@ -24,34 +24,74 @@ function BoardWrite() {
         location: '서울',
         files: []
     })
+
+
+
     // 파일업로드
+
+    const handleUploadedFiles = (files) => {
+        setUploadedFiles(files);
+    };
+
+
     useEffect(() => {
         setIsFileUploadDisabled(false); // 처음 로드할 때 파일 업로드 활성화
     }, []);
 
+    // 파일업로드
     const handleFileChange = (files) => {
-        setFormData((prevData) => ({
-            ...prevData,
-            files: files // 선택한 파일들을 files 상태에 업데이트
-        }));
-        // 업로드된 파일 상태 업데이트
-        setUploadedFiles(files); // 업로드된 파일 상태 업데이트
-        // 파일 업로드 비활성화 설정
-        setIsFileUploadDisabled(true);
-        console.log('Selected Files:', files); // 파일 선택 시 로그 출력
+        const newFiles = [];
+
+        const onUpload = (file, reader) => {
+            reader.onload = () => {
+                const fileObject = {
+                    name: file.name,
+                    uuid: file.name,
+                    origin: file.type,
+                    src: reader.result,
+                    file_path: '/images/' + file.name
+                };
+                newFiles.push(fileObject);
+
+                if (newFiles.length === files.length) {
+                    setUploadedFiles(newFiles); // Update uploadedFiles state
+                    setFormData((prevData) => ({
+                        ...prevData,
+                        files: newFiles
+                    }));
+                    setIsFileUploadDisabled(false); // 파일 업로드 활성화
+                    console.log('Selected Files:', newFiles); // 파일 선택 시 로그 출력
+                }
+            };
+        };
+
+        for (let i = 0; i < files.length; i++) {
+            const file = files[i];
+            const reader = new FileReader();
+            onUpload(file, reader);
+            reader.readAsDataURL(file);
+        }
     };
+
+    // 작성(입력)
     const handleChange = (e) => {
         const { name, value } = e.target;
 
         // 실행하려는 로직
-        if (name === 'boardCno') {
-            if (value === '2') {
-                setIsFileUploadDisabled(true); // 비활성화
-            } else {
-                setIsFileUploadDisabled(false);
-            }
-        }
+        // if (name === 'boardCno') {
+        //     if (value === '2') {
+        //         setIsFileUploadDisabled(true); // 비활성화
+        //     } else {
+        //         setIsFileUploadDisabled(false);
+        //     }
+        // }
 
+        if (name === 'title') {
+            setFormData((prevData) => ({
+                ...prevData,
+                title: value
+            }));
+        }
         // 폼 데이터 업데이트
         setFormData((prevData) => ({
             ...prevData,
@@ -102,49 +142,6 @@ function BoardWrite() {
             alert("게시글 작성 오류")
         }
     }
-
-    // BoardFileUpload.js 
-    // const BoardFileUpload = ({ isDisabled }) => {
-    //     const [uploadedFiles, setUploadedFiles] = useState([]);
-    
-    //     const onUpload = (e) => {
-    //         if (!isDisabled) {
-    //             const files = e.target.files;
-    //             if (files) {
-    //                 const newFiles = [];
-    
-    //                 for (let i = 0; i < files.length; i++) {
-    //                     const file = files[i];
-    //                     const reader = new FileReader();
-    
-    //                     reader.onload = () => {
-    //                         // 원하는 속성을 가진 새로운 파일 객체를 생성
-    //                         const fileObject = {
-    //                             name: file.name, // 파일 이름
-    //                             uuid: file.name, // UUID 함수를 사용하여 고유한 값 생성
-    //                             origin: file.type, // 원본 파일 타입
-    //                             src: reader.result,
-    //                             file_path: '/images/' + file.name // 파일 경로 (서버에 저장될 경로)
-    //                             // ... 다른 필요한 정보들을 여기에 추가할 수 있음
-    //                         };
-    //                         newFiles.push(fileObject);
-    
-    //                         if (newFiles.length === files.length) {
-    //                             setUploadedFiles((prevUploadedFiles) => {
-    //                                 return [...prevUploadedFiles, ...newFiles];
-    //                             });
-    //                             console.log('Uploaded Files:', [...uploadedFiles, ...newFiles]); // 새로 업로드된 파일들을 콘솔에 출력
-    //                         }
-    //                     };
-    
-    //                     reader.readAsDataURL(file);
-    //                 }
-    //             }
-    //         }
-    //     };
-
-
-
 
 
     return (
@@ -229,15 +226,41 @@ function BoardWrite() {
                         <div className='underline' />
 
                         <Col md={6} className='d-flex justify-content-end'>
-                            <BoardUploadFile isDisabled={isFileUploadDisabled}
+                          
+                            {/* <BoardFileUpload isDisabled={isFileUploadDisabled}
                                 uploadedFiles={uploadedFiles} // 업로드된 파일 상태 전달
-                                setUploadedFiles={setUploadedFiles}>
+                                // setUploadedFiles={setUploadedFiles }>
+                                setUploadedFiles={handleUploadedFiles }>
                                 <input
                                     type="file"
                                     multiple // 여러 파일을 선택할 수 있도록 multiple 속성 추가
                                     onChange={(e) => handleFileChange(e.target.files)} // 파일 변경 시 handleFileChange 함수 호출
                                 />
-                            </BoardUploadFile>
+                            </BoardFileUpload> */}
+
+
+                            <div className="input-group mb-3">
+                                <input
+                                    accept="image/*"
+                                    multiple
+                                    type="file"
+                                    onChange={(e) => handleFileChange(e.target.files)}
+                                    className="form-control"
+                                    id="inputGroupFile02"
+                                    disabled={isFileUploadDisabled}
+                                />
+                                <label className="input-group-text" htmlFor="inputGroupFile02">
+                                    업로드
+                                </label>
+                            </div>
+
+                            {/* 업로드된 이미지 미리보기 */}
+                            {uploadedFiles.map((fileObject, index) => (
+                                <div key={index} style={{ display: 'flex', flexDirection: 'column', alignItems: 'center', marginBottom: '10px' }}>
+                                    <img src={fileObject.src} alt='FileUpload' style={{ width: '200px', height: '200px', objectFit: 'cover', marginBottom: '5px' }} />
+                                    <p>{fileObject.name}</p>
+                                </div>
+                            ))}
                         </Col>
 
 
@@ -254,7 +277,7 @@ function BoardWrite() {
                             />
                         </Form.Group>
                         <Col className="d-flex justify-content-end">
-                            <Button variant="primary" type="button" id="btn-save" onClick={handleSubmit}>
+                            <Button variant="primary" type="submit" id="btn-save" onClick={handleSubmit}>
                                 등록
                             </Button>
                         </Col>
@@ -264,5 +287,5 @@ function BoardWrite() {
         </>
     );
 }
-// }
+{/* } */ }
 export default BoardWrite;
